completing the login flow

+++ client/src/components/ShoppingList.jsx
@@ -1,16 +1,29 @@
 import PropTypes from 'prop-types';
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import { connect } from 'react-redux';
 import { CSSTransition, TransitionGroup } from 'react-transition-group';
-import { Button, Container, ListGroup, ListGroupItem } from 'reactstrap';
+import { Alert, Button, Container, ListGroup, ListGroupItem } from 'reactstrap';
 import { deleteItem, getItems } from '../actions/itemActions';
 
 const ShoppingList = props => {
-  const { item: { items }, getItems, deleteItem } = props;
+  const { item: { items }, auth: { isAuthenticated, user }, getItems, deleteItem } = props;
+
+  const [alert, setalert] = useState(false);
 
   useEffect(() => {
     getItems();
   }, [getItems]);
+
+  useEffect(() => {
+    if(isAuthenticated) {
+      setalert(true);
+      setTimeout(() => {
+        setalert(false);
+      }, 3000);
+    } else {
+      setalert(false);
+    }
+  }, [isAuthenticated]);
 
   const onRemoveItem = id => {
     deleteItem(id);
@@ -25,11 +38,12 @@ const ShoppingList = props => {
               <ListGroupItem>
                 <Button
                   className="rmv-btn"
-                  color="danger"
+                  color={isAuthenticated ? "danger" : "info"}
                   size="sm"
                   onClick={() => onRemoveItem(_id)}
+                  disabled={!isAuthenticated}
                 >
-                  &times;
+                  {isAuthenticated ? <>&#10005;</> : <>&#10140;</>}
                 </Button>
                 {name}
               </ListGroupItem>
@@ -37,17 +51,26 @@ const ShoppingList = props => {
           ))}
         </TransitionGroup>
       </ListGroup>
+      {
+        alert && (
+          <Alert className="mt-4">
+            {`Welcome ${user.name}! Now you can manage the list.`}
+          </Alert>
+        )
+      }
     </Container>
   );
 };
 
 ShoppingList.propTypes = {
   getItems: PropTypes.func.isRequired,
-  item: PropTypes.object.isRequired
+  item: PropTypes.object.isRequired,
+  auth: PropTypes.object.isRequired
 };
 
 const mapStateToProps = state => ({
-  item: state.item
+  item: state.item,
+  auth: state.auth
 });
 
 export default connect(mapStateToProps, { getItems, deleteItem })(ShoppingList);
